Move existing data loading test to new test module
